pyproject.toml
[project]
name = "elections"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "psycopg[binary]>=3.2.0",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
]
